refactor(python): lift shared infra into _paycore so protocols don't depend on each other
Address lgalabru: 'protocols should not depend on each other.' x402 was
importing SolanaRpc, the replay Store, the network-blockhash check, and the
v0-wire detector from protocols/mpp. Move that shared infrastructure into
_paycore (the analog of the rust 'core' crate): errors, rpc, store,
network_check, and a new transaction.py (is_v0_wire_bytes). Both x402 and mpp
now depend only on _paycore; neither imports the other (verified: zero
protocols.x402<->protocols.mpp references).

No wire/behavior change: 700 tests pass, pyright/ruff clean, rust interop stays
green for charge and x402-exact.
