+++ python/src/pay_kit/__init__.py
@@ -26,6 +26,7 @@ from pay_kit._paycore.currency import Currency
 from pay_kit._paycore.network import Network
 from pay_kit._paycore.protocol import Protocol
 from pay_kit._paycore.stablecoin import Stablecoin
+from pay_kit._paycore.store import FileReplayStore, MemoryStore, Store
 from pay_kit.config import (
     Config,
     MppConfig,
@@ -55,7 +56,6 @@ from pay_kit.payment import Payment
 from pay_kit.price import Price
 from pay_kit.pricing import Pricing
 from pay_kit.protocols.mpp.core.expires import days, hours, minutes, seconds, weeks
-from pay_kit.protocols.mpp.core.store import FileReplayStore, MemoryStore, Store
 from pay_kit.signer import LocalSigner, Signer
 
 __all__ = [
