+++ python/tests/test_cross_route_replay.py
@@ -9,10 +9,10 @@ from __future__ import annotations
 
 import pytest
 
+from pay_kit._paycore.errors import PaymentError
+from pay_kit._paycore.store import MemoryStore
 from pay_kit.protocols.mpp.core.base64url import encode_json
 from pay_kit.protocols.mpp.core.challenge import compute_challenge_id
-from pay_kit.protocols.mpp.core.errors import PaymentError
-from pay_kit.protocols.mpp.core.store import MemoryStore
 from pay_kit.protocols.mpp.core.types import ChallengeEcho, PaymentCredential
 from pay_kit.protocols.mpp.intents.charge import ChargeRequest
 from pay_kit.protocols.mpp.server.charge import Config, Mpp
